requirements.txt
Flask==0.12
google-cloud==0.23.0
google-api-python-client==1.6.2
